policy statement moved to a separate file
